tool/pyproject.toml
[project]
name = "tool"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "pillow>=10.0",
    "requests>=2.31",
    "numpy>=1.26"
]

[build-system]
requires = ["pillow"]
build-backend = "setuptools.build_meta"
